chore: extract servicelogger to pkg
