+++ app/artifact-cas/internal/service/resource.go
@@ -20,7 +20,7 @@ import (
 
 	v1 "github.com/chainloop-dev/chainloop/app/artifact-cas/api/cas/v1"
 	backend "github.com/chainloop-dev/chainloop/internal/blobmanager"
-	sl "github.com/chainloop-dev/chainloop/internal/servicelogger"
+	sl "github.com/chainloop-dev/chainloop/pkg/servicelogger"
 	"github.com/go-kratos/kratos/v2/errors"
 )
