+++ app/controlplane/internal/biz/attestation.go
@@ -22,7 +22,7 @@ import (
 	"io"
 
 	"github.com/chainloop-dev/chainloop/internal/attestation"
-	"github.com/chainloop-dev/chainloop/internal/servicelogger"
+	"github.com/chainloop-dev/chainloop/pkg/servicelogger"
 	"github.com/go-kratos/kratos/v2/log"
 
 	cr_v1 "github.com/google/go-containerregistry/pkg/v1"
