+++ app/controlplane/plugins/sdk/v1/fanout.go
@@ -29,7 +29,7 @@ import (
 
 	schemaapi "github.com/chainloop-dev/chainloop/app/controlplane/api/workflowcontract/v1"
 	"github.com/chainloop-dev/chainloop/internal/attestation/renderer/chainloop"
-	"github.com/chainloop-dev/chainloop/internal/servicelogger"
+	"github.com/chainloop-dev/chainloop/pkg/servicelogger"
 	"github.com/go-kratos/kratos/v2/log"
 	intoto "github.com/in-toto/attestation/go/v1"
 	"github.com/invopop/jsonschema"
