+++ internal/credentials/azurekv/keyvault.go
@@ -29,7 +29,7 @@ import (
 	"github.com/chainloop-dev/chainloop/internal/credentials"
 	"github.com/google/uuid"
 
-	"github.com/chainloop-dev/chainloop/internal/servicelogger"
+	"github.com/chainloop-dev/chainloop/pkg/servicelogger"
 	"github.com/go-kratos/kratos/v2/log"
 )
