+++ internal/credentials/gcp/secretmanager.go
@@ -28,7 +28,7 @@ import (
 	"google.golang.org/api/option"
 
 	"github.com/chainloop-dev/chainloop/internal/credentials"
-	"github.com/chainloop-dev/chainloop/internal/servicelogger"
+	"github.com/chainloop-dev/chainloop/pkg/servicelogger"
 	"github.com/docker/distribution/uuid"
 	"github.com/go-kratos/kratos/v2/log"
 	"github.com/googleapis/gax-go/v2"
