+++ internal/credentials/vault/keyval.go
@@ -26,7 +26,7 @@ import (
 	"time"
 
 	"github.com/chainloop-dev/chainloop/internal/credentials"
-	"github.com/chainloop-dev/chainloop/internal/servicelogger"
+	"github.com/chainloop-dev/chainloop/pkg/servicelogger"
 	"github.com/docker/distribution/uuid"
 	"github.com/go-kratos/kratos/v2/log"
 	vault "github.com/hashicorp/vault/api"
